+++ huskarl/CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] - 2026-04-28
+
+### Improvements
+
+- Extra claims are no longer wrapped in an option.
+
+### Changes
+
+- Bump huskarl-core to 0.3
+- Bump huskarl-crypto-native to 0.4
+- Bump huskarl-crypto-webcrypto to 0.3 
+
 ## [0.3.0]
 
 ### Changes